// src/domain/models/Theme.tsx
export type Theme = {
    lang: string;
    dir: string;
    dataThemeMode: string;
    dataMenuStyles: string;
    dataNavLayout: string;
    dataHeaderStyles: string;
    dataVerticalStyle: string;
    StylebodyBg: string;
    StyleDarkBg: string;
    toggled: string;
    dataNavStyle: string;
    horStyle: string;
    dataPageStyle: string;
    dataWidth: string;
    dataMenuPosition: string;
    dataHeaderPosition: string;
    loader: string;
    iconOverlay: string;
    colorPrimaryRgb: string;
    bodyBg1: string;
    bodyBg2: string;
    darkBg: string;
    inputBorder: string;
    bgImg: string;
    iconText: string;
    body: {
        class: string;
    };
}